Changed .discard to .remove
Also when the user sends a snap and RandoBot is not on the list, it will
now send a snap saying it must be added to the friend list, so they're
not improperly put into the list. Attempted to fix some indentation
issues as well, as spacing by four wasn't working for some? Commented
lines are for if/when your bot gets blocked. Remove the first three
comments of code, connect to your old bot and let it copy users into
lists. Then, comment out those lines and uncomment the rest, to re-add
all your users!